Decompose addAnswerToAnswersListAsync in SurveyScreen to insertAnswerAsync in answers helper function
