+++ src/__tests__/SurveyScreen.test.tsx
@@ -8,7 +8,9 @@ import {
 import waitForExpect from "wait-for-expect";
 
 import SurveyScreen, { SurveyScreenProps } from "../SurveyScreen";
+import { getAnswerEntity } from "../entities/AnswerEntity";
 import { PingEntity } from "../entities/PingEntity";
+import * as HelperAnswers from "../helpers/answers";
 import { QuestionType } from "../helpers/helpers";
 import * as HelperPings from "../helpers/pings";
 import { QuestionsList } from "../helpers/types";
@@ -55,8 +57,27 @@ const NEXT_BUTTON_A11YLABEL = "Next question";
 function mockDatabaseRelatedFunction() {
   // https://stackoverflow.com/a/56565849/2603230
   jest
-    .spyOn(SurveyScreen.prototype, "addAnswerToAnswersListAsync")
-    .mockImplementation(async () => {});
+    .spyOn(HelperAnswers, "insertAnswerAsync")
+    .mockImplementation(
+      async ({
+        ping,
+        question,
+        realQuestionId,
+        preferNotToAnswer,
+        nextWithoutOption,
+        data,
+        lastUpdateDate,
+      }) => {
+        const answer = getAnswerEntity(question.type);
+        answer.ping = ping;
+        answer.questionId = realQuestionId;
+        answer.preferNotToAnswer = preferNotToAnswer;
+        answer.nextWithoutOption = nextWithoutOption;
+        answer.data = data;
+        answer.lastUpdateDate = lastUpdateDate;
+        return answer;
+      },
+    );
 
   jest
     .spyOn(HelperPings, "addEndTimeToPingAsync")
